bootstrap + main objects
Loading at the bootstrap Polyglot.js which fetches the current language
of the browser. Also creating the AppRouter, the AppController
(displays the right views for the different urls triggered by the
AppRouter), the DataController which will make ajax calls, the
AppLayoutView which is the main view for the whole application. Adding
also the LoadingView for the spinner, as well as a LoginView, which
will contain the form to sign-in or sign-up.