moving under /account + userTypeChoiceView
Moving the classes linked with the account process under /account
folder. Adding a new class when user wants to create a new account,
which is only displaying 2 buttons to choose the type of user he/she
will be.
Also now possible to ‘log-in’ now, meaning to obtain an access-token
from Auth0 if the user already has an account created.