new club page + geocomplete lib
Adding the page allowing a new manager to create his club by entering
the name and the address. The address is using the google autocomplete
but thru the jquery.geocomplete lib.
Adding the new url that goes it (/profile/club).
But the confirm button does not call any Rest service yet.
Making a lot of css for the photo page and the club page.