new agency page
Same as for the club page, a new Agency page has been added to enter
the name of address of the agency, which is displayed only when a new
user is created and the rd-btn Agency is choosen.
Fixing some css issues, as well as changing the main color of the app.